feat: move mutation dialog to its own file and act upon onClose
